+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ New Features
 
 Fixed
 ^^^^^
+- Fixed :gh:`139` BM3D tensor format grayscale (:gh:`140` by `Matthieu Terris`_) - 23/12/2023
 - Fixed :gh:`136` noise additive model for DecomposablePhysics (:gh:`138` by `Matthieu Terris`_) - 22/12/2023